+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixed
 - Fixed RESP3 double parsing returning positive `INF` for `-inf` payloads (#1716)
 - Fixed `client_info` connection parameter being ignored (#1722)
+- Fixed `Stream::write()`/`read()` leaving a dead connection when a host error handler throws exception (#1725)
 
 ## v3.6.0 (2026-08-14)
 ### Added